--- a/src/Nest/CommonAbstractions/Infer/Field/ToStringExpressionVisitor.cs
+++ b/src/Nest/CommonAbstractions/Infer/Field/ToStringExpressionVisitor.cs
@@ -9,12 +9,9 @@ using System.Text;
 
 namespace Nest
 {
-	internal class FieldExpressionVisitor : ExpressionVisitor
+	internal class ToStringExpressionVisitor : ExpressionVisitor
 	{
-		private readonly IConnectionSettingsValues _settings;
 		private readonly Stack<string> _stack = new Stack<string>();
-
-		public FieldExpressionVisitor(IConnectionSettingsValues settings) => _settings = settings;
 
 		public string Resolve(Expression expression, bool toLastToken = false)
 		{
@@ -29,21 +26,7 @@ namespace Nest
 				.ToString();
 		}
 
-		public string Resolve(MemberInfo info)
-		{
-			if (info == null) return null;
-
-			var name = info.Name;
-
-			if (_settings.PropertyMappings.TryGetValue(info, out var propertyMapping))
-				return propertyMapping.Name;
-
-			var att = ElasticsearchPropertyAttributeBase.From(info);
-			if (att != null && !att.Name.IsNullOrEmpty())
-				return att.Name;
-
-			return _settings.PropertyMappingProvider?.CreatePropertyMapping(info)?.Name ?? _settings.DefaultFieldNameInferrer(name);
-		}
+		public string Resolve(MemberInfo info) => info == null ? null : info.Name;
 
 		protected override Expression VisitMember(MemberExpression expression)
 		{
@@ -91,10 +74,17 @@ namespace Nest
 		private static void VisitConstantOrVariable(MethodCallExpression methodCall, Stack<string> stack)
 		{
 			var lastArg = methodCall.Arguments.Last();
-			var value = lastArg is ConstantExpression constantExpression
-				? constantExpression.Value.ToString()
-				: Expression.Lambda(lastArg).Compile().DynamicInvoke().ToString();
-			stack.Push(value);
+			if (lastArg is ConstantExpression constantExpression)
+			{
+				stack.Push(constantExpression.Value.ToString());
+				return;
+			}
+			if (lastArg is MemberExpression memberExpression)
+			{
+				stack.Push(memberExpression.Member.Name);
+				return;
+			}
+			stack.Push(lastArg.ToString());
 		}
 
 		private static bool IsLinqOperator(MethodInfo methodInfo)