@Mpdreamz 2019-04-15
Tried locally to reuse the visit of the `ToStringVisitor` and it creating a second stack for the fieldname to resolve so it doesn't need to visit the expression as well.

Don't see any real benefit and it adds quite some complexity so resetting locally.

Benchmark for reference


|                   Method |       Jit | Runtime |     Mean |      Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------- |-------- |---------:|-----------:|----------:|------------:|------------:|------------:|--------------------:|
|       'Boxed Expression' | LegacyJit |     Clr | 5.914 ns |  7.5240 ns | 0.4124 ns |      0.0004 |           - |           - |                   - |
|     'Unboxed Expression' | LegacyJit |     Clr | 3.445 ns |  3.2429 ns | 0.1778 ns |      0.0003 |           - |           - |                   - |
|         'Field Resolver' | LegacyJit |     Clr | 5.022 ns |  1.4263 ns | 0.0782 ns |      0.0004 |           - |           - |                   - |
| 'Field Resolver Unboxed' | LegacyJit |     Clr | 3.657 ns |  4.1828 ns | 0.2293 ns |      0.0003 |           - |           - |                   - |
|       'Boxed Expression' |    RyuJit |     Clr | 6.113 ns |  0.0929 ns | 0.0051 ns |      0.0004 |           - |           - |                   - |
|     'Unboxed Expression' |    RyuJit |     Clr | 3.317 ns |  0.1476 ns | 0.0081 ns |      0.0003 |           - |           - |                   - |
|         'Field Resolver' |    RyuJit |     Clr | 6.347 ns | 31.5901 ns | 1.7316 ns |      0.0004 |           - |           - |                   - |
| 'Field Resolver Unboxed' |    RyuJit |     Clr | 3.754 ns |  0.0714 ns | 0.0039 ns |      0.0003 |           - |           - |                   - |
|       'Boxed Expression' |    RyuJit |    Core | 5.315 ns |  0.9369 ns | 0.0514 ns |      0.0003 |           - |           - |                 2 B |
|     'Unboxed Expression' |    RyuJit |    Core | 3.855 ns |  7.7585 ns | 0.4253 ns |      0.0003 |           - |           - |                 2 B |
|         'Field Resolver' |    RyuJit |    Core | 6.263 ns |  0.8671 ns | 0.0475 ns |      0.0004 |           - |           - |                 2 B |
| 'Field Resolver Unboxed' |    RyuJit |    Core | 4.779 ns |  0.2023 ns | 0.0111 ns |      0.0003 |           - |           - |                 2 B |