@Mpdreamz 2019-04-16
cherry-picked to master